app/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-multiselectfield==0.1.13
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
